requirements.txt
Django==2.2
django-debug-toolbar==1.11
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.2
PyMySQL==0.9.3
pytz==2019.1
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0
Whoosh==2.7.4
